test_requirements.txt
# Production Readiness Testing Requirements
# Install these dependencies to run the comprehensive testing suite

requests>=2.28.0
psutil>=5.9.0
Pillow>=9.0.0  # For image processing tests
Flask>=2.3.0   # Required for importing app modules
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0

# Optional but recommended for enhanced testing
selenium>=4.0.0  # For browser automation tests (if needed)
pytest>=7.0.0    # Alternative test runner
coverage>=7.0.0  # For code coverage analysis
